kotlinx module refactor package to `com.fleeksoft.io.kotlinx`
