+++ kotlinx-io/src/com/fleeksoft/io/kotlinx/SourceExt.kt
@@ -1,5 +1,6 @@
-package com.fleeksoft.io
+package com.fleeksoft.io.kotlinx
 
+import com.fleeksoft.io.InputStream
 import kotlinx.io.RawSource
 import kotlinx.io.Source
 import kotlinx.io.buffered
