+++ kotlinx-io/src/com/fleeksoft/io/kotlinx/SourceInputStream.kt
@@ -1,5 +1,6 @@
-package com.fleeksoft.io
+package com.fleeksoft.io.kotlinx
 
+import com.fleeksoft.io.InputStream
 import kotlinx.io.InternalIoApi
 import kotlinx.io.Source
 import kotlinx.io.readByteArray
